insurance-chatbot/backend/requirements.txt
langchain
langchain-community
torch
transformers
sentence-transformers
faiss-cpu
unstructured
gradio
accelerate
bitsandbytes
langchain-huggingface
fastapi
uvicorn